+++ app/views/questions/_question_content.erb
@@ -1,0 +1,21 @@
+<h1 class="text-3xl font-extrabold mb-4"><%= question.title %></h1>
+
+<div class="mb-6">
+    <div class="space-y-4">
+        <div id="voting-section">
+            <%= render 'options_with_results', question: question, current_user: current_user %>
+        </div>
+
+        <div class="flex items-center space-x-3 mt-8 pt-4 border-t border-gray-200">
+        <img src="<%= question.user.google_image_url %>" alt="<%= question.user.name %>" class="w-10 h-10 rounded-full">
+        <span class="font-md"><%= t('.creater')%>: <%= @question.user.name %></span>
+
+        <% if current_user == question.user %>
+            <div class="ml-auto flex items-center space-x-2">
+                <%= link_to t('.edit'), edit_question_path(question), data: { turbo_frame: "question_#{question.id}" }, class: "text-gray-500 hover:text-gray-900" %>
+                <%= button_to t('.destroy'), question_path(question), method: :delete, form: { data: { turbo_confirm: '本当に削除しますか？', turbo_frame: '_top'}}, class: "text-red-500 hover:text-red-700" %>
+            </div>
+        <% end %>
+        </div>
+    </div>
+</div>
