--- app/views/questions/_show.html.erb
@@ -1,35 +1,0 @@
-<div id="question_#{@question.id}">
-  <div class="container mx-auto p-4 sm:p-6 lg:p-8">
-    <div class="flex flex-col lg:flex-row lg:space-x-8">
-
-      <div class="flex-grow lg:w-2/3 mb-8 lg:mb-0">
-        <div class="bg-white shadow-lg rounded-lg p-6 mb-6 border border-gray-200">
-          <h1 class="text-3xl font-extrabold mb-4"><%= @question.title %></h1>
-
-          <div class="mb-6">
-            <div class="space-y-4">
-                <div id="voting-section">
-                  <%= render 'options_with_results', question: @question, current_user: current_user %>
-                </div>
-
-              <div class="flex items-center space-x-3 mt-8 pt-4 border-t border-gray-200">
-                <img src="<%= @question.user.google_image_url %>" alt="<%= @question.user.name %>" class="w-10 h-10 rounded-full">
-                <span class="font-md"><%= t('.creater')%>: <%= @question.user.name %></span>
-
-                <% if current_user == @question.user %>
-                  <div class="ml-auto flex items-center space-x-2">
-                    <%= link_to t('.edit'), edit_question_path(@question), class: "text-gray-500 hover:text-gray-900", data: { turbo_frame: "question_#{@question.id}" } %>
-                    <%= button_to t('.destroy'), question_path(@question), method: :delete, class: "text-red-500 hover:text-red-700", data: { turbo_confirm: '本当に削除しますか？' } %>
-                  </div>
-                <% end %>
-              </div>
-            </div>
-          </div>
-        </div>
-      <div class="text-center mt-4">
-        <%= link_to t('.return_index'), questions_path, class: "text-blue-500 hover:underline" %>
-      </div>
-
-    </div>
-  </div>
-</div>
