+++ app/views/questions/edit.html.erb
@@ -1,5 +1,12 @@
-<%= turbo_frame_tag "question_#{@question.id}" do %>
-    <h1 class="text-center text-lg mt-4"><%= t('.title')%></h1>
-    <%= render 'form', question_form: @question_form %>
-<% end %>
+<div class="container mx-auto p-4 sm:p-6 lg:p-8">
+  <div class="flex flex-col lg:flex-row lg:space-x-8">
 
+    <div class="flex-grow lg:w-2/3 mb-8 lg:mb-0">
+      <div class="bg-white shadow-lg rounded-lg p-6 mb-6 border border-gray-200">
+        <%= turbo_frame_tag "question_#{@question.id}" do %>
+        <%= render 'edit_form', question: @question %>
+        <% end %>
+      </div>
+    </div>
+  </div>
+</div>
