+++ app/views/questions/show.html.erb
@@ -1,3 +1,15 @@
-<%= turbo_frame_tag "question_#{@question.id}" do %>
-  <%= render "show", question: @question %>
-<% end %>
+<div class="container mx-auto p-4 sm:p-6 lg:p-8">
+  <div class="flex flex-col lg:flex-row lg:space-x-8">
+
+    <div class="flex-grow lg:w-2/3 mb-8 lg:mb-0">
+      <div class="bg-white shadow-lg rounded-lg p-6 mb-6 border border-gray-200">
+      <%= turbo_frame_tag "question_#{@question.id}" do %>
+        <%= render 'question_content', question: @question, current_user: current_user %>
+      <% end %>
+      </div>
+    <div class="text-center mt-4">
+      <%= link_to t('.return_index'), questions_path, class: "text-blue-500 hover:underline" %>
+    </div>
+
+  </div>
+</div>
