+++ app/views/questions/update.turbo_stream.erb
@@ -1,4 +1,4 @@
 <%= turbo_stream.replace "question_#{@question.id}" do %>
-  <%= render "questions/show", question: @question_form.question %>
+  <%= render "question_content", question: @question_form.question%>
 <% end %>
 <%= turbo_stream_flash %>
