+++ config/locales/views/ja.yml
@@ -33,12 +33,16 @@ ja:
     index:
       no_questions: お題はありません
     show:
-      creater: 作成者
       return_index: お題一覧に戻る
-      edit: 編集
-      destroy: 削除
     options_with_results:
       choose: どちらか選択してください
       vote_login_prompt: "投票するにはログインしてください"
       re-vote: "投票をやり直す"
       already_voted: "このお題は投票済みです"
+    question_content:
+      creater: 作成者
+      edit: 編集
+      destroy: 削除
+    edit_form:
+      creater: 作成者
+      cancel: キャンセル
